Método compute_train_performance adicionado a classe ClassificadorBinario
