+++ pycomp/ml/trainer.py
@@ -13,7 +13,13 @@ Sumário
 """
 
 # Importando bibliotecas
+import logging
+import time
+import pandas as pd
+from pycomp.log.log_config import log_config
 from sklearn.model_selection import RandomizedSearchCV
+from sklearn.model_selection import cross_val_score, cross_val_predict
+from sklearn.metrics import roc_auc_score
 
 
 """
@@ -84,25 +90,100 @@ class ClassificadorBinario:
         """
 
         # Iterando sobre os modelos presentes no dicionário de classificadores
-        for model_name, model_info in set_classifiers.items():
-            # Definindo chave do classificador para o dicionário classifiers_info
-            clf_key = model_name + approach
-            logger.debug(f'Treinando modelo {clf_key}')
-            model = model_info['model']
+        try:
+            for model_name, model_info in set_classifiers.items():
+                # Definindo chave do classificador para o dicionário classifiers_info
+                clf_key = model_name + approach
+                logger.debug(f'Treinando modelo {clf_key}')
+                model = model_info['model']
 
-            # Criando dicionário vazio para armazenar dados do modelo
-            self.classifiers_info[clf_key] = {}
+                # Criando dicionário vazio para armazenar dados do modelo
+                self.classifiers_info[clf_key] = {}
 
-            # Validando aplicação da busca aleatória pelos melhores hiperparâmetros
-            if random_search:
-                params = model_info['params']
-                rnd_search = RandomizedSearchCV(model, params, scoring=scoring, cv=cv, verbose=verbose,
-                                                random_state=42, n_jobs=n_jobs)
-                rnd_search.fit(X, y)
+                # Validando aplicação da busca aleatória pelos melhores hiperparâmetros
+                try:
+                    if random_search:
+                        params = model_info['params']
+                        rnd_search = RandomizedSearchCV(model, params, scoring=scoring, cv=cv, verbose=verbose,
+                                                        random_state=42, n_jobs=n_jobs)
+                        logger.debug('Aplicando RandomizedSearchCV')
+                        rnd_search.fit(X, y)
 
-                # Salvando melhor modelo no atributo classifiers_info
-                self.classifiers_info[clf_key]['estimator'] = rnd_search.best_estimator_
-            else:
-                # Treinando modelo sem busca e salvando no atirbuto
-                self.classifiers_info[clf_key]['estimator'] = model.fit(X, y)
+                        # Salvando melhor modelo no atributo classifiers_info
+                        self.classifiers_info[clf_key]['estimator'] = rnd_search.best_estimator_
+                    else:
+                        # Treinando modelo sem busca e salvando no atirbuto
+                        self.classifiers_info[clf_key]['estimator'] = model.fit(X, y)
+                    logger.info(f'Modelo {clf_key} treinado com sucesso')
+                except ValueError as e:
+                    logger.error(f'Erro ao treinar o modelo {clf_key}. Exception lançada: {e}')
+                    continue
+        except AttributeError as e:
+            logger.error('Dicionário de classificador(es) preparado de forma incorreta. Utilize {model_name: {estimator: estimator, params: params}}')
+            logger.warning(f'Treinamento do(s) modelo(s) não realizado')
 
+    def compute_train_performance(self, model_name, estimator, X, y, cv=5):
+        """
+        Método responsável por aplicar validação cruzada para retornar as principais métricas de avaliação
+        de um modelo de classificação. Na prática, esse método é chamado por um outro método em uma camada
+        superior da classe para medição de performance em treino e em teste
+
+        Parâmetros
+        ----------
+        :param model_name: chave identificadora do modelo contida no atributo self.classifiers_info [type: string]
+        :param estimator: estimator do modelo a ser avaliado [type: object]
+        :param X: conjunto de features do modelo contido nos dados de treino [type: np.array]
+        :param y: array contendo a variável resposta dos dados de trieno do modelo [type: np.array]
+        :param cv: K-folds utiliados na validação cruzada [type: int, default: 5]
+
+        Retorno
+        -------
+        :return train_performance: DataFrame contendo as métricas calculadas usando validação cruzada [type: pd.DataFrame]
+
+        Aplicação
+        ---------
+        # Instanciando e treinando modelo
+        trainer = ClassificadorBinario()
+        trainer.fit(model, X_train, y_train)
+        train_performance = trainer.compute_train_performance(model_name, estimator, X_train, y_train)
+        """
+
+        # Computando métricas utilizando validação cruzada
+        logger.debug(f'Computando métricas do modelo {model_name} utilizando validação cruzada com {cv} K-folds')
+        try:
+            t0 = time.time()
+            accuracy = cross_val_score(estimator, X, y, cv=cv, scoring='accuracy').mean()
+            precision = cross_val_score(estimator, X, y, cv=cv, scoring='precision').mean()
+            recall = cross_val_score(estimator, X, y, cv=cv, scoring='recall').mean()
+            f1 = cross_val_score(estimator, X, y, cv=cv, scoring='f1').mean()
+
+            # Probabilidades para o cálculo da AUC
+            try:
+                y_scores = cross_val_predict(estimator, X, y, cv=cv, method='decision_function')
+            except:
+                # Modelos baseados em árvore não possuem o método decision_function() mas sim o predict_proba()
+                y_probas = cross_val_predict(estimator, X, y, cv=cv, method='predict_proba')
+                y_scores = y_probas[:, 1]
+            auc = roc_auc_score(y, y_scores)
+
+            # Salvando métricas no atributo self.classifiers_info
+            self.classifiers_info[model_name]['train_scores'] = y_scores
+
+            # Criando DataFrame com o resultado obtido
+            t1 = time.time()
+            delta_time = t1 - t0
+            train_performance = {}
+            train_performance['model'] = model_name
+            train_performance['approach'] = f'Treino {cv} K-folds'
+            train_performance['acc'] = round(accuracy, 4)
+            train_performance['precision'] = round(precision, 4)
+            train_performance['recall'] = round(recall, 4)
+            train_performance['f1'] = round(f1, 4)
+            train_performance['auc'] = round(auc, 4)
+            train_performance['total_time'] = round(delta_time, 3)
+
+            logger.info(f'Métricas computadas com sucesso nos dados de treino em {round(delta_time, 3)} segundos')
+        except Exception as e:
+            logger.error(f'Erro ao computar as métricas. Exception lançada: {e}')
+
+        return pd.DataFrame(train_performance, index=train_performance.keys()).reset_index(drop=True).loc[:0, :]
